+++ Net/src/StreamSocket.cpp
@@ -136,7 +136,7 @@ int StreamSocket::sendBytes(const void* buffer, int length, int flags)
 
 int StreamSocket::sendBytes(FIFOBuffer& fifoBuf)
 {
-	int ret = impl()->sendBytes(&fifoBuf.buffer()[0], fifoBuf.used());
+	int ret = impl()->sendBytes(&fifoBuf.buffer()[0], (int) fifoBuf.used());
 	if (ret > 0) fifoBuf.drain(ret);
 	return ret;
 }
@@ -150,7 +150,7 @@ int StreamSocket::receiveBytes(void* buffer, int length, int flags)
 
 int StreamSocket::receiveBytes(FIFOBuffer& fifoBuf)
 {
-	int ret = impl()->receiveBytes(fifoBuf.next(), fifoBuf.available());
+	int ret = impl()->receiveBytes(fifoBuf.next(), (int) fifoBuf.available());
 	if (ret > 0) fifoBuf.advance(ret);
 	return ret;
 }
